Moved admin routes handling to `AdminService`
